python_uart_reader_docker/app/requirements.txt
Flask==1.1.1
pySerial
